Design about section
I finished designing the about section. This section is responsive
across all devices. I used a lot of media queries to tightly control
the layout, because I wanted make the about section fit in 100% of any
viewport. I realized that this is not entirely desirable for mobile
devices. Therefor, for smaller screens, the section will need to be
scrolled.